smart_habit_tracker/requirements.txt
acme==1.21.0
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
aniso8601==9.0.1
apturl==0.5.2
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.1.0
Babel==2.12.1
bcrypt==3.2.0
bidict==0.22.1
billiard==4.2.1
blinker==1.6.2
Brlapi==0.8.3
celery==5.4.0
certbot==1.21.0
certbot-nginx==1.21.0
certifi==2024.2.2
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.4
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
command-not-found==0.3
ConfigArgParse==1.5.3
configobj==5.0.6
coverage==7.8.0
cryptography==3.4.8
cupshelpers==1.0
DateTime==5.2
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
dotenv==0.9.9
duplicity==0.8.21
exceptiongroup==1.2.2
fasteners==0.14.1
filelock==3.6.0
Flask==2.3.2
Flask-Cors==4.0.0
Flask-Moment==1.0.5
Flask-RESTful==0.3.10
Flask-SocketIO==5.3.4
Flask-WTF==1.1.1
frozenlist==1.4.0
future==0.18.2
gevent==22.10.2
greenlet==2.0.2
gunicorn==20.1.0
httplib2==0.20.2
idna==3.7
importlib-metadata==4.6.4
iniconfig==2.1.0
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
josepy==1.10.0
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
keyring==23.5.0
kombu==5.4.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.3
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.0.4
mysql-connector-python==8.1.0
netifaces==0.11.0
numpy==1.25.1
oauthlib==3.2.0
olefile==0.46
packaging==24.2
paho-mqtt==2.1.0
paramiko==2.9.3
parsedatetime==2.6
pexpect==4.8.0
pika==1.3.2
Pillow==9.0.1
pipenv==11.9.0
platformdirs==2.5.1
pluggy==1.5.0
prompt_toolkit==3.0.48
protobuf==4.21.12
psycopg2-binary==2.9.7
ptyprocess==0.7.0
pycairo==1.20.1
pycryptodome==3.19.0
pycups==2.0.1
PyGObject==3.42.1
PyICU==2.8.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyMySQL==1.1.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyproj==3.6.0
pyRFC3339==1.1
pyrsistent==0.18.1
pytest==8.2.2
pytest-asyncio==0.24.0
pytest-cov==5.0.0
pytest-django==4.11.1
pytest-docker-fixtures==1.3.19
pytest-mqtt==0.4.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-engineio==4.5.1
python-socketio==5.8.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
redis==5.2.0
referencing==0.29.1
reportlab==3.6.8
requests==2.32.2
requests-toolbelt==0.9.1
rpds-py==0.8.10
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
secure-smtplib==0.1.1
shapely==2.0.1
six==1.16.0
sqlparse==0.5.3
ssh-import-id==5.11
sshtunnel==0.4.0
supervisor==4.2.1
sympy==1.12
systemd-python==234
texttable==1.6.4
tomli==2.0.2
twilio==8.8.0
typing_extensions==4.12.2
tzdata==2025.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.2.1
usb-creator==0.3.7
vine==5.1.0
virtualenv==20.13.0+ds
virtualenv-clone==0.3.0
wadllib==1.3.6
wcwidth==0.2.13
websocket-client==1.2.3
Werkzeug==2.3.6
WTForms==3.0.1
xdg==5
xkit==0.0.0
yarl==1.9.2
zipp==1.0.0
zope.component==4.3.0
zope.event==4.4
zope.hookable==5.1.0
zope.interface==5.4.0
